requirements.txt
asyncio==3.4.3
discord.py>=1.3.3
colorama==0.4.3
GitPython==2.1.15
google-api-core==1.17.0
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-translate==2.0.1
googleapis-common-protos==1.51.0
pathlib2==2.3.5
Pillow>=7
PyNaCl==1.3.0
python-dateutil==2.8.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
subprocess32==3.5.4
youtube_dl>=2020.6.6
